requirements.txt
Django==4.1.4
numpy== 1.21.5
mysql-connector-python==8.0.31
pandas==1.4.4
pandas-profiling==3.5.0
plotly==5.11.0
python-dotenv==0.21.0
django-cors-headers==3.13.0
scikit-learn ==1.0.2
mlflow == 2.7.1
scikit-optimize== 0.9.0
PyMySQL==1.0.2
xgboost==1.7.4
pyreadstat==1.2.1
openpyxl==3.1.2
google-auth==2.15.0
celery== 5.2.7
django-celery-results==2.5.0
django-memcached==0.1.2
pylibmc==1.6.3
cryptography==41.0.2
mysqlclient==2.1.1
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-django==0.41b0